Correct handling of context data
This is a big commit, but doesn't really change the main logic.

* src/Interpreters/IndexContextInfo.h (IndexContextInfo): New data
structure to store the index information that will be stored in the
context.
* src/Processors/QueryPlan/ReadFromMergeTree.cpp (BuildIndexes):
Doesn't modify the context anymore because it hasn't filtered all the
marks yet.
*
src/Storages/MergeTree/MergeTreeDataSelectExecutor.cpp (filterPartsByPrimaryKeyAndSkipInd)
: Use the new structure to store index information in the
context. This change simplifies the access to the context as the part
information is set in parallel in a temporal vector (lock-free) and
the ranges are copied after passing multiple filters, which is good to
avoid reading granules without matches.
* src/Functions/hasTokenIndex.h: Use the new context structure with
the index information. Add code to handle filtered/skipped marks more
efficiently.